New feature for task number 3.1.3

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/RoleDao.java
@@ -14,7 +14,7 @@ import java.util.Optional;
 public interface RoleDao {
     List<Role> listRoles();
 
-    Optional<Role> findById(int id);
+    Optional<Role> findById(Long id);
 
     Optional<Role> findByName(String name);
