+++ src/main/java/ru/kata/spring/boot_security/demo/services/RoleService.java
@@ -16,7 +16,7 @@ public interface RoleService {
 
     Optional<Role> findByName(String name);
 
-    Optional<Role> findById(int id);
+    Optional<Role> findById(Long id);
 
     Role getDefaultRole();
 }
