+++ src/main/java/ru/kata/spring/boot_security/demo/services/RoleServiceImpl.java
@@ -32,7 +32,7 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Optional<Role> findById(int id) {
+    public Optional<Role> findById(Long id) {
         return roleDao.findById(id);
     }
